FILE: extracts Filters interface to a separate file
