+++ libs/spirit-islander/shared/util/src/lib/functions/filters.interface.ts
@@ -1,0 +1,6 @@
+import { Expansion } from '../types';
+
+export interface Filters<TName extends string> {
+  expansions?: readonly Expansion[];
+  names?: readonly TName[];
+}
