+++ libs/spirit-islander/shared/util/src/lib/functions/get-adversaries.ts
@@ -1,6 +1,7 @@
 import { ADVERSARIES } from '../data';
 import { Adversary, AdversaryName } from '../types';
-import { Filters, getOptions } from './get-options';
+import { Filters } from './filters.interface';
+import { getOptions } from './get-options';
 
 export function getAdversaries(filters: Filters<AdversaryName> = {}) {
   return getOptions<AdversaryName, Adversary>(ADVERSARIES, filters);
