+++ libs/spirit-islander/shared/util/src/lib/functions/get-boards.ts
@@ -1,6 +1,7 @@
 import { BOARDS } from '../data';
 import { BalancedBoardName, Board } from '../types';
-import { Filters, getOptions } from './get-options';
+import { Filters } from './filters.interface';
+import { getOptions } from './get-options';
 
 export function getBoards(filters: Filters<BalancedBoardName> = {}) {
   return getOptions<BalancedBoardName, Board>(BOARDS, filters);
