+++ libs/spirit-islander/shared/util/src/lib/functions/get-options.ts
@@ -1,9 +1,5 @@
-import { Expansion, ExpansionOption } from '../types';
-
-export interface Filters<TName extends string> {
-  expansions?: readonly Expansion[];
-  names?: readonly TName[];
-}
+import { ExpansionOption } from '../types';
+import { Filters } from './filters.interface';
 
 export function getOptions<
   TName extends string,
