+++ libs/spirit-islander/shared/util/src/lib/functions/index.ts
@@ -11,3 +11,4 @@ export * from './get-players';
 export * from './get-scenarios';
 export * from './get-spirits';
 export * from './group-spirits';
+export * from './filters.interface';
